+++ src/schema/resolvers/roleResolvers.js
@@ -76,31 +76,15 @@ export default {
         throw new ValidationError('bad-request');
       }
     },
-    deleteRole: async (
-      root,
-      { id },
-      { models: { Role, DefaultPermission, Permission }, ValidationError },
-    ) => {
+    deleteRole: async (root, { id }, { models: { Role }, ValidationError }) => {
       try {
-        // delete permission with role_id
-        const permissionWithFeatureID = Permission.query().where('role_id', id);
-        if (permissionWithFeatureID) {
-          await Permission.query()
-            .delete()
-            .where('role_id', id);
-        }
-
-        // delete default permission with role_id
-        const defaultPermissionWithFeatureID = DefaultPermission.query().where(
-          'role_id',
-          id,
-        );
-        if (defaultPermissionWithFeatureID) {
-          await DefaultPermission.query()
-            .delete()
-            .where('role_id', id);
-        }
         const role = await Role.query().findById(id);
+        if (!role) throw new ValidationError('Role not found');
+        // delete defaultpermission with role id
+        // delete permission with role id
+        role.deleteAllRelationship();
+
+        // delete role
         await Role.query().deleteById(id);
         return role;
       } catch (e) {
